Improve enum-of-byte handling
Expressions where comparisons of enum-of-byte are made are always cast to int comparisons by the C# compiler. This causes SQL queries to do unnecessary casts of tinyint column like CAST ([column] as int)

Counteract that by detecting comparisons of byte enums with other byte enums or int/byte constants and see if they both can be cast to byte.